feat: Update dependencies in pdm.lock and pyproject.toml
This commit updates the dependencies in the `pdm.lock` and `pyproject.toml` files. It includes upgrading the `lock_version` in `pdm.lock` to `4.5.0` and updating the `requires_python` version in `pdm.lock` and `pyproject.toml` to `3.12.*`. Additionally, it adds new dependencies such as `eval-type-backport`, `importlib-metadata`, `importlib-resources`, `exceptiongroup`, `typing-extensions`, `trio`, and `httpx` in `pyproject.toml`. These updates ensure that the application has access to the latest stable versions of the required packages and enhances its functionality.